Move type checking time classes to stub
